+++ BeeWare/package.json
@@ -20,7 +20,6 @@
     "react-dom": "^16.13.0",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.0",
     "typescript": "3.8.3"
   },
   "scripts": {
@@ -45,7 +44,8 @@
     ]
   },
   "devDependencies": {
-    "@capacitor/cli": "2.4.2"
+    "@capacitor/cli": "2.4.2",
+    "react-scripts": "3.4.0"
   },
   "description": "An Ionic project"
 }
