+++ BeeWare/public/assets/icon/favicon.ico
