+++ BeeWare/src/data/Riddles.tsx
