+++ BeeWare/src/models/Riddle.tsx
