+++ BeeWare/src/pages/Riddle.tsx
