--- beeware/BeeWare/capacitor.config.json
@@ -1,13 +1,0 @@
-{
-  "appId": "io.ionic.starter",
-  "appName": "BeeWare",
-  "bundledWebRuntime": false,
-  "npmClient": "npm",
-  "webDir": "build",
-  "plugins": {
-    "SplashScreen": {
-      "launchShowDuration": 0
-    }
-  },
-  "cordova": {}
-}
