--- beeware/BeeWare/public/manifest.json
@@ -1,20 +1,0 @@
-{
-    "short_name": "Ionic App",
-    "name": "My Ionic App",
-    "icons": [{
-            "src": "assets/icon/favicon.ico",
-            "sizes": "64x64 32x32 24x24 16x16",
-            "type": "image/x-icon"
-        },
-        {
-            "src": "assets/icon/icon.png",
-            "type": "image/png",
-            "sizes": "512x512",
-            "purpose": "maskable"
-        }
-    ],
-    "start_url": ".",
-    "display": "standalone",
-    "theme_color": "#ffffff",
-    "background_color": "#ffffff"
-}
